// app/interfaces/MainComponent/IBlogBoxComponent.ts
export interface IBlogBox {
    _id: string;
    id?: string;
    pagetitle?: string;
    title?: string;
    image?: string;
    description?: string;
    }
    
    export interface IBlogBoxInput {
        id?: string;
        pagetitle?: string;
        title?: string;
        image?: string;
        description?: string;
    }